Модуляция
Разделение кода из main.js на отдельные ES2015-модули
