+++ js/utils/message.js
@@ -1,0 +1,1 @@
+// Модуль вывода сообщений
