+++ js/utils/post.js
@@ -1,0 +1,7 @@
+// *** МОДУЛЬ ВЫВОДА ДАННЫХ
+
+/* eslint-disable no-console */
+
+import {objectsNearby} from './data';
+
+console.log(objectsNearby);
